requirements.txt
torch
torchvision
streamlit
pillow
scikit-learn
joblib
pandas
requests